+++ tests/test_ranking.py
@@ -60,3 +60,12 @@ def test_result_negative_score():
     except ValueError as e:
         assert str(e) == "Scores must be non-negative"
 
+def test_result_league_points_home_win():
+    team1 = Team("Lions")
+    team2 = Team("Grouches")
+    result = Result(team1, 3, team2, 1)
+    team1_expected_points = LeaguePoints(team1, 3)
+    team2_expected_points = LeaguePoints(team2, 0)
+
+    assert result.league_points().contains(team1_expected_points)
+    assert result.league_points().contains(team2_expected_points)